better than before, not done

+++ Shelly/Shelly/main.cpp
@@ -9,42 +9,52 @@
  *      they refer to the latest halted or stopped process.
  */
 
+
 #include "Shelly.h"
 
 Shelly shelly;
 
-void handle_SIGSEGV(int signum)
-{
+void handle_SIGSEGV(int signum) {
     std::cout << "SEG_FAULT:" << signum << std::endl;
-    return;
+
 }
 
-void handle_keyC(int signum)
-{
-    Process* process = shelly.getPHandler()->getLastFrontProcess();
-    pid_t pid = process->getPid();
-    kill(pid, SIGINT);
-    std::cout << "Stopped Process - pid:" << pid;
-    return;
+void handle_keyC(int signum) {
+    std::cout << "\n pressed C" << std::endl;
+    if (shelly.getPHandler()->getLastBackProcess() != NULL) {
+        Process* process = shelly.getPHandler()->getLastFrontProcess();
+        pid_t pid = process->getPid();
+        kill(pid, SIGINT);
+        std::cout << "Stopped Process - pid:" << pid;
+    } else {
+        std::string in;
+
+        std::cout << "\n\n"; //4 debug
+        std::cout << "Wirklich Beenden? Y or n\n>>";
+        std::cin>>in;
+
+        if (in == "Y" || in == "y" || in == "j" || in == "J") {
+            std::cin.ignore(187, '\n');
+            std::cout << "\nHauste" << std::endl;
+        }
+    }
 }
 
-void handle_keyZ(int signum)
-{
+void handle_keyZ(int signum) {
+    std::cout << "\n pressed Z" << std::endl;
     Process* process = shelly.getPHandler()->getLastFrontProcess();
     pid_t pid = process->getPid();
     kill(pid, SIGTSTP);
     std::cout << "Halted process - pid:" << pid;
-    return;
+
 }
 
-void handle_SIGCHLD(int signum)
-{
+void handle_SIGCHLD(int signum) {
     pid_t l_pid;
     int status;
     l_pid = waitpid(-1, &status, WNOHANG);
     if (l_pid == -1);
-    else
-    {
+    else {
         Process* process = shelly.getPHandler()->getProcess(l_pid);
         process->changeStatus(Process::ProcessStatus::endet);
         std::cout << "\nProcess " << l_pid << " terminated" << std::endl;
@@ -52,20 +62,16 @@ void handle_SIGCHLD(int signum)
     }
 }
 
-int main()
-{
+int main() {
     signal(SIGSEGV, handle_SIGSEGV);
     signal(SIGINT, handle_keyC);
     signal(SIGSTOP, handle_keyZ);
     signal(SIGCHLD, handle_SIGCHLD);
 
 
-    try
-    {
+    try {
         while (shelly.execute());
-    }
-    catch (...)
-    {
+    } catch (...) {
         std::cout << "Catch Something Mainly" << std::endl;
     }
